pyproject.toml
[project]
name = "spec-review"
version = "0.1.0"
description = "MEP Specification Review Tool for California K-12 DSA Projects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "python-docx>=1.1.0",
    "rich>=13.0.0",
    "tiktoken>=0.7.0",
]

[project.scripts]
spec-review = "src.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
